third_party/WebKit/LayoutTests/external/wpt/uievents/auxclick/auxclick_event-manual-expected.txt

This is a testharness.js-based test.
FAIL auxclick event sequence received. assert_equals: detail attribute of auxclick should be the click count. expected 2 but got 1
PASS PointerEvent Automation
PASS First auxclick should have detail=1 indicating the fire click
FAIL Second auxclick should have detail=2 indicating the fire click assert_equals: detail attribute of auxclick should be the click count. expected 2 but got 1
Harness: the test ran to completion.
